KS-2-1. fix

+++ src/main/kotlin/lesson_2/l2_task1.kt
@@ -5,7 +5,7 @@ fun main() {
     val rating2 = 4
     val rating3 = 3
     val rating4 = 5
-    val quantity = 4.0f
-    val average = (rating1 + rating2 + rating3 + rating4) / quantity
+    val quantity = 4
+    val average = (rating1 + rating2 + rating3 + rating4).toFloat() / quantity
     println(String.format("%.2f", average))
 }
